# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail CRM API",
        "description": "API для управления розничной CRM-системой: товары, склады, заказы, поставки, перемещения, клиенты.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/supplies": {
            "get": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Список поставок",
                "description": "Возвращает поставки с пагинацией и фильтрацией по складу и дате.",
                "operationId": "34cd493d00dc7571cc7a6a5f913dd6bc",
                "parameters": [
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Фильтр по складу",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата от (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата до (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кол-во на странице (по умолчанию 15)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пагинированный список поставок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Supply"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Создать поставку",
                "description": "Создаёт поставку на склад. Увеличивает остатки товаров и записывает движения.",
                "operationId": "8519da10dfd698919213f167529033fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSupplyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Поставка создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Supply"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Список перемещений",
                "description": "Возвращает перемещения с пагинацией и фильтрацией по складам-отправителю/получателю и дате.",
                "operationId": "5f1c0157d61b1f15afbbb54382ca5e1f",
                "parameters": [
                    {
                        "name": "from_warehouse_id",
                        "in": "query",
                        "description": "Фильтр по складу-отправителю",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "to_warehouse_id",
                        "in": "query",
                        "description": "Фильтр по складу-получателю",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата от (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата до (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кол-во на странице (по умолчанию 15)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пагинированный список перемещений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transfer"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Создать перемещение",
                "description": "Перемещает товары между складами. Проверяет наличие на складе-отправителе. Создаёт два движения остатков (списание и приход).",
                "operationId": "f9336a57ea8062efd2bbd119f12b170f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Перемещение создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Transfer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации / недостаточно товара на складе-отправителе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnavailableProducts"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SupplyItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Футболка белая"
                    },
                    "count": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "Supply": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "warehouse_name": {
                        "type": "string",
                        "example": "Основной склад"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SupplyItem"
                        }
                    }
                },
                "type": "object"
            },
            "TransferItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Футболка белая"
                    },
                    "count": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "Transfer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "from_warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "from_warehouse_name": {
                        "type": "string",
                        "example": "Основной склад"
                    },
                    "to_warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "to_warehouse_name": {
                        "type": "string",
                        "example": "Филиал №2"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransferItem"
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreSupplyItem": {
                "required": [
                    "product_id",
                    "count"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "count": {
                        "type": "integer",
                        "example": 100,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "StoreSupplyRequest": {
                "required": [
                    "warehouse_id",
                    "items"
                ],
                "properties": {
                    "warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreSupplyItem"
                        }
                    }
                },
                "type": "object"
            },
            "StoreTransferItem": {
                "required": [
                    "product_id",
                    "count"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "count": {
                        "type": "integer",
                        "example": 20,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "StoreTransferRequest": {
                "required": [
                    "from_warehouse_id",
                    "to_warehouse_id",
                    "items"
                ],
                "properties": {
                    "from_warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "to_warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreTransferItem"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnavailableProducts": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Some products are not available in sufficient quantity."
                    },
                    "unavailable_products": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "product_id": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "integer"
                                },
                                "available": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}